requirements/base.txt
aiomysql==0.1.1
aioredis==2.0.1
aiosmtplib==1.1.6
anyio==3.6.1
async-timeout==4.0.2
bcrypt==3.2.2
bidict==0.22.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
databases==0.6.1
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.0
fastapi==0.85.0
fastapi-pagination==0.10.0
greenlet==1.1.3
gTTS==2.2.4
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
makefun==1.15.0
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
passlib==1.7.4
pycparser==2.21
pydantic==1.10.2
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
python-engineio==4.3.4
python-multipart==0.0.5
python-socketio==5.7.0
pytube==15.0.0
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
wrapt==1.14.1
